+++ providers/aws/services/ec2/ec2_ebs_snapshots_encrypted/ec2_ebs_snapshots_encrypted.metadata.json
@@ -48,7 +48,7 @@
   "ResourceIdTemplate": "arn:partition:service:region:account-id:resource-id",
   "ResourceType": "AwsIamAccessAnalyzer",
   "Risk": "Risk associated.",
-  "ServiceName": "iam",
+  "ServiceName": "ec2",
   "Severity": "low",
   "SubServiceName": "accessanalyzer",
   "Tags": {
